+++ pajbot/modules/duel.py
@@ -37,6 +37,15 @@ class DuelModule(BaseModule):
             constraints={"min_value": 0, "max_value": 1000000},
         ),
         ModuleSetting(
+            key="min_pot",
+            label="The minimum amount of points that must be used for a duel",
+            type="number",
+            required=True,
+            placeholder="",
+            default=0,
+            constraints={"min_value": 0, "max_value": 1000000},
+        ),
+        ModuleSetting(
             key="message_won",
             label="Winner message | Available arguments: {winner}, {loser}",
             type="text",
@@ -82,7 +91,11 @@ class DuelModule(BaseModule):
             constraints={"min_value": 0, "max_value": 240},
         ),
         ModuleSetting(
-            key="show_on_clr", label="Show duels on the clr overlay", type="boolean", required=True, default=True
+            key="show_on_clr",
+            label="Show duels on the clr overlay",
+            type="boolean",
+            required=True,
+            default=True,
         ),
         ModuleSetting(
             key="max_duel_age",
@@ -162,6 +175,19 @@ class DuelModule(BaseModule):
             return False
 
         max_pot = self.settings["max_pot"]
+        assert isinstance(max_pot, int)
+        min_pot = self.settings["min_pot"]
+        assert isinstance(min_pot, int)
+
+        msg_id = rest["args"]["msg_id"]
+        assert isinstance(msg_id, str)
+
+        if min_pot > max_pot:
+            bot.reply(
+                msg_id,
+                f"Duel module is misconfigured, the minimum duel amount ({min_pot}) is higher than the maximum duel amount ({max_pot})",
+            )
+            return False
 
         msg_split = message.split()
         input = msg_split[0]
@@ -183,6 +209,13 @@ class DuelModule(BaseModule):
                         duel_price = max_pot
                 except ValueError:
                     pass
+
+            if duel_price < min_pot:
+                bot.whisper(
+                    source,
+                    f"You must duel for at least {min_pot} points.",
+                )
+                return False
 
             if source.id in self.duel_requests:
                 currently_duelling = User.find_by_id(db_session, self.duel_requests[source.id])
@@ -385,7 +418,7 @@ class DuelModule(BaseModule):
                 bot.whisper(source, "You have no duel request or duel target. Type !duel USERNAME POT to duel someone!")
 
     @staticmethod
-    def get_duel_stats(bot: Bot, source: User, **rest: Any) -> None:
+    def get_duel_stats(bot: Bot, source: User, **rest: Any) -> bool:
         """
         Whispers the users duel winratio to the user
         """
@@ -397,6 +430,8 @@ class DuelModule(BaseModule):
             source,
             f"duels: {source.duel_stats.duels_total} winrate: {source.duel_stats.winrate:.2f}% streak: {source.duel_stats.current_streak} profit: {source.duel_stats.profit}",
         )
+
+        return True
 
     def _cancel_expired_duels(self) -> None:
         if self.bot is None: